nit: make MetricDataPoint an `int` instead of `double`
We're never actually using doubles currently, so doesn't seem worth using them here for now

+++ tracer/src/Datadog.Trace/Telemetry/DTOs/MetricSeries.cs
@@ -59,7 +59,7 @@ internal class MetricSeries : List<MetricDataPoint>
                 }
 
                 var timestamp = pointArrayValues[0].Value<long>();
-                var value = pointArrayValues[1].Value<double>();
+                var value = pointArrayValues[1].Value<int>();
 
                 data.Add(new MetricDataPoint(timestamp, value));
             }
